requirements.txt
Flask==3.0.0
flask-cors==4.0.0
boto3==1.26.0
python-dotenv==1.0.0
